feat: コマンド一覧をmain.goに移動
